+++ async/src/test/kotlin/lt/saltyjuice/dragas/chatty/v3/async/mock/route/MockRouter.kt
@@ -3,7 +3,7 @@ package lt.saltyjuice.dragas.chatty.v3.async.mock.route
 import kotlinx.coroutines.experimental.channels.SendChannel
 import lt.saltyjuice.dragas.chatty.v3.async.route.AsyncRouter
 
-class MockRouter(channel: SendChannel<Int>) : AsyncRouter<Int, Int>(channel)
+class MockRouter(channel: SendChannel<Float>) : AsyncRouter<Int, Float>(channel)
 {
     override fun builder(): MockRoute.Builder
     {
